pyproject.toml
[project]
name = "proxygen"
version = "0.1.0"
description = "Proxy Generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "httpx>=0.28.1",
    "logging>=0.4.9.6",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "uvicorn>=0.38.0",
]
